swnpuerto/requirements/base.txt
Django==1.11.3
Pillow==2.8.2
Unipath==1.1
django-model-utils==3.0.0
djangorestframework==3.6.3
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
